+++ providers/capmox/standard/capmox_test.go
@@ -1,0 +1,16 @@
+package standard
+
+import (
+	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
+
+	"github.com/giantswarm/cluster-test-suites/v6/internal/common"
+)
+
+var _ = XDescribe("Common tests", func() {
+	cfg := common.NewTestConfigWithDefaults()
+	// No autoscaling on-prem
+	cfg.AutoScalingSupported = false
+	// Disabled until https://github.com/giantswarm/roadmap/issues/1037
+	cfg.ExternalDnsSupported = false
+	common.Run(cfg)
+})